--- MEIA/src/main/java/com/meia_2020/meia/Bitacora_Usuario.java
@@ -1,17 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.meia_2020.meia;
-
-import com.meia_2020.meia.models.Usuario;
-
-/**
- *
- * @author roche
- */
-public class Bitacora_Usuario {
-          
-   
-}
