+++ MEIA/src/main/java/com/meia_2020/meia/LoginForm.java
@@ -1,6 +1,7 @@
 package com.meia_2020.meia;
 
 
+import com.meia_2020.meia.models.Data;
 import com.meia_2020.meia.models.Usuario;
 import java.awt.List;
 import java.io.BufferedReader;
