+++ MEIA/src/main/java/com/meia_2020/meia/models/Bitacora_Usuario.java
@@ -3,12 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.meia_2020.meia;
+package com.meia_2020.meia.models;
+
+import com.meia_2020.meia.models.Usuario;
 
 /**
  *
  * @author roche
  */
-public class ArchivoUsuario {
-    
+public class Bitacora_Usuario {
+          
+   
 }
