+++ MEIA/src/main/java/com/meia_2020/meia/models/Data.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.meia_2020.meia;
+package com.meia_2020.meia.models;
 
 import com.meia_2020.meia.models.Usuario;
