+++ MEIA/src/main/java/com/meia_2020/meia/models/bitacora_backup.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package com.meia_2020.meia;
+package com.meia_2020.meia.models;
 
 import java.time.LocalDate;
